remove cloud function from sds codebase
